+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - support return of null in variable replacement (#513)
 - allow for variable use in $pick picklists: `{{ $pick ask-variable? $value: data }}` (#537)
+- pre request in intellij format is executed before variable replacer (#534)
 
 ## [6.6.3]  (2023-07-25)